+++ src/_packages/core.index.ts
@@ -3,6 +3,7 @@
 
 import { BOOK_LANGUAGE_VERSION, PROMPTBOOK_ENGINE_VERSION } from '../version';
 import { createAgentModelRequirements } from '../book-2.0/agent-source/createAgentModelRequirements';
+import { createAgentModelRequirementsWithCommitments } from '../book-2.0/agent-source/createAgentModelRequirementsWithCommitments';
 import { parseAgentSource } from '../book-2.0/agent-source/parseAgentSource';
 import { parseParameters } from '../book-2.0/agent-source/parseParameters';
 import { isValidBook } from '../book-2.0/agent-source/string_book';
@@ -179,6 +180,7 @@ export { BOOK_LANGUAGE_VERSION, PROMPTBOOK_ENGINE_VERSION };
 
 // Note: Entities of the `@promptbook/core`
 export { createAgentModelRequirements };
+export { createAgentModelRequirementsWithCommitments };
 export { parseAgentSource };
 export { parseParameters };
 export { isValidBook };